# lib/api/openapi.yaml
openapi: "3.1.0"
info:
  title: "HAY2010 Stock API"
  version: "1.0.0"
  description: "API REST pour la gestion des partenaires, documents, stock et affaires"
servers:
  - url: "/api/v1"
    description: "Base path for API v1"
security:
  - ApiKeyAuth: []
paths:
...
  /produits/{id}/niveaux-stock:
    get:
      operationId: getProductStockLevels
      summary: "Lister les niveaux de stock d'un produit"
      tags:
        - Produits
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: "Liste paginée des niveaux de stock"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockLevelPaginatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /entrepots/{id}/niveaux-stock:
    get:
      operationId: getWarehouseStockLevels
      summary: "Lister les niveaux de stock d'un entrepôt"
      tags:
        - Entrepots
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: "Liste paginée des niveaux de stock"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockLevelPaginatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /niveaux-stock:
    get:
      operationId: listStockLevels
      summary: "Lister les niveaux de stock"
      tags:
        - NiveauxStock
      security:
        - ApiKeyAuth: []
      parameters:
        - name: produit
          in: query
          schema:
            type: integer
          description: "Filtrer par ID produit"
        - name: entrepot
          in: query
          schema:
            type: integer
          description: "Filtrer par ID entrepôt"
        - name: sort
          in: query
          schema:
            type: string
            enum: [id_stock, id_produit, id_entrepot, quantite_en_stock, quantite_reservee, quantite_commandee, date_dernier_mouvement, date_creation, date_modification]
            default: date_creation
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: "Liste paginée des niveaux de stock"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullStockLevelPaginatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createStockLevel
      summary: "Créer un niveau de stock"
      tags:
        - NiveauxStock
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockLevelCreate"
      responses:
        "201":
          description: "Niveau de stock créé"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockLevel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: "Produit ou entrepôt non trouvé"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: "Combinaison produit-entrepôt déjà existante"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /niveaux-stock/{id}:
    get:
      operationId: getStockLevelById
      summary: "Obtenir un niveau de stock par ID"
      tags:
        - NiveauxStock
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Détails du niveau de stock"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullStockLevel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateStockLevel
      summary: "Mettre à jour un niveau de stock"
      tags:
        - NiveauxStock
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockLevelUpdate"
      responses:
        "200":
          description: "Niveau de stock mis à jour"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockLevel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Combinaison produit-entrepôt déjà existante"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteStockLevel
      summary: "Supprimer un niveau de stock"
      tags:
        - NiveauxStock
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: "Niveau de stock supprimé"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "API key au format hay2010_sk_live_xxxxxxxxxxxxxxxx"
  responses:
    Unauthorized:
      description: "Clé API invalide ou manquante"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: "Requête invalide"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: "Ressource non trouvée"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
...
    StockLevel:
      type: object
      properties:
        id_stock:
          type: integer
        id_produit:
          type: integer
        id_entrepot:
          type: integer
        quantite_en_stock:
          type: number
        quantite_reservee:
          type: number
        quantite_commandee:
          type: number
        date_dernier_mouvement:
          type: string
          format: date-time
          nullable: true
        type_dernier_mouvement:
          type: string
          nullable: true
        date_creation:
          type: string
          format: date-time
        date_modification:
          type: string
          format: date-time
    StockLevelCreate:
      type: object
      required:
        - id_produit
        - id_entrepot
      properties:
        id_produit:
          type: integer
        id_entrepot:
          type: integer
        quantite_en_stock:
          type: number
        quantite_reservee:
          type: number
        quantite_commandee:
          type: number
        date_dernier_mouvement:
          type: string
          nullable: true
        type_dernier_mouvement:
          type: string
          maxLength: 50
          nullable: true
    StockLevelUpdate:
      type: object
      description: "Tous les champs sont optionnels pour la mise à jour partielle"
      properties:
        id_produit:
          type: integer
        id_entrepot:
          type: integer
        quantite_en_stock:
          type: number
        quantite_reservee:
          type: number
        quantite_commandee:
          type: number
        date_dernier_mouvement:
          type: string
          nullable: true
        type_dernier_mouvement:
          type: string
          maxLength: 50
          nullable: true
    StockLevelPaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/StockLevel"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    FullStockLevel:
      type: object
      allOf:
        - $ref: "#/components/schemas/StockLevel"
        - type: object
          properties:
            produit:
              type: object
              properties:
                id_produit:
                  type: integer
                code_produit:
                  type: string
                nom_produit:
                  type: string
            entrepot:
              type: object
              properties:
                id_entrepot:
                  type: integer
                code_entrepot:
                  type: string
                nom_entrepot:
                  type: string
    FullStockLevelPaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FullStockLevel"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties:
            type: object
            properties:
              message:
                type: string
              code:
                type: string
        timestamp:
          type: string
          format: date-time